requirements.txt
streamlit
pypdf
sentence-transformers
faiss-cpu
groq
numpy
tqdm
